Allow user to override the Code Point Limit required by SnakeYaml
With this patch, a user may set the system property
'maxYamlCodePoints' in order to override the default 3MiB limit
configured in the org.yaml.snakeyaml package by default. This limit
was implemented to prevent certain Denial-of-Service (DOS) attacks,
but users should be given the opportunity to override this value for
valid configurations which exceed the limit, such as the Redfish
OpenAPI specification (developed by DMTF), which weighs in at 4.9MiB.

This patch was tested to work with openapi-generator-cli v6.3.0.

Signed-off-by: Ethan D. Twardy <ethan.twardy@gmail.com>