Correctly resolve externalRefs that point to the root path
Previously, if OpenAPI v3 document A loads document B, and B refers
to schemas in document A, swagger-parser would not resolve references
to schemas in A as internal refs, even though they refer to the root
OpenAPI v3 document (document A). This would result in the creation of
duplicate schemas for those schemas in document A. Now, we resolve these
"external" refs as internal refs, preventing the duplication of these
schemas by the ExternalRefProcessor.